+++ test/emoji.parse-fail.js
@@ -1,0 +1,16 @@
+'use strict';
+
+import test from 'ava';
+
+var characters = require('./fixtures/emoji-characters.js');
+var emoji = require('../parsers/emoji.js');
+
+characters.forEach(function (e) {
+  test('parse fail on emoji with non-emoji a' + e + 'a', function (t) {
+    t.plan(1);
+
+    var result = emoji.parse(`a${e}a`);
+
+    t.false(result);
+  });
+});
